vhost-iommu: Implement IOMMU accelerator for VHOST devices
Since VHOST allows to handle fast path vring traffic in kernel
it does make sense to handle IOVA translation in kernel as well.
This way we avoid unnecessary host kernel and QEMU communication.

For the sake of code review VHOST IOMMU is implemented as separate
driver. However, it is possible to combine it with VIRTIO IOMMU
and then consider VHOST as an extra mode.

Note that VIRTIO_IOMMU_F_PROBE is not supported yet.

Signed-off-by: Tomasz Nowicki <tn@semihalf.com>