vhost: Add infrastructure to register VHOST device as behind the VHOST IOMMU
For a proper translation requests handling VHOST IOMMU expects
VHOST devices to be registered which means VHOST devices are placed
behind the VHOST IOMMU in hierarchy.

Add extra callback to VirtioBusClass to allow VIRTIO devices to
be registered as behind the IOMMU. Also, implement mentioned callback
for VIRTIO PCI bus in first place.

Signed-off-by: Tomasz Nowicki <tn@semihalf.com>