hw/arm/sysbus-fdt: helpers for clock-mmio node generation
Some passthrough'ed devices depend on clock nodes. Those need to be
generated in the guest device tree. This patch introduces some helpers
to instantiate and build a clock-mmio node from information retrieved in
the host device tree.

- fdt_clock_mmio_realize() instantiates clock-mmio device and sets its
  properties like: related vfio device, name, index
- fdt_build_clock_mmio_node() calls fdt_clock_mmio_realize() and builds
  a guest clock node

fdt_clock_mmio_realize() will become static as soon as it gets used. A
dummy pre-declaration is needed for compilation of this patch.

Signed-off-by: Tomasz Nowicki <tn@semihalf.com>