vhost-vfio: add Virtio VFIO device
Implement the new Virtio device for host<->guest communication using
the Virtio queues. This is VHOST-based configuration medium for devices
which are directly assigned to virtual machine (via VFIO) and needs other
host devices help to work properly like clock, PHY etc. Most of the work
is done in a vhost kernel driver.

Signed-off-by: Tomasz Nowicki <tn@semihalf.com>