hw/arm/sysbus-fdt: helpers for host node properties handling
Passthrough'ed devices rely on its DT properties and those need to be
handled in a different ways depending on the situation. This patch
introduces some helpers to create per-device tables with specified
property name and actions that supposed to be taken when copying from
host node.

- copy_node_prop() traverses host node properties and lookup table
  property to find out what actions should be taken: copy as is, warn
  about occurrence, ignore or simply reject (which results in exit)
- copy_host_node_prop() finds the VFIO device host DT node and calls
  copy_node_prop() to handle host to guest DT properties modification

copy_host_node_prop() will become static as soon as it gets used. A
dummy pre-declaration is needed for compilation of this patch.

Signed-off-by: Tomasz Nowicki <tn@semihalf.com>
Change-Id: I5f92e2fb1e9809a10ac83f012e6ac2851267b5da