hw/arm/sysbus-fdt: helpers for virtio-clk node generation
Some passthrough'ed devices depend on clock nodes. Those need to be
generated in the guest device tree. This patch introduces helpers
to instantiate and build a virtio-clk node based on information retrieved
from the host device tree.

- fdt_virtio_clk_realize() instantiates virtio-clk device and sets its
  properties like: virtio dev type, related vfio device, name, index
- fdt_build_virtio_clk_node() calls fdt_clock_mmio_realize() and builds
  a guest clock node

fdt_virtio_clk_realize() will become static as soon as it gets used. A
dummy pre-declaration is needed for compilation of this patch.

Signed-off-by: Tomasz Nowicki <tn@semihalf.com>